+++ tests/test_fits.py
@@ -592,8 +592,15 @@ def test_ndarray_validation(tmp_path):
 
     # But raise an error when casting is disabled
     with pytest.raises(ValidationError, match="Array datatype 'float64' is not compatible with 'float32'"):
-        with FitsModel(file_path, strict_validation=True, cast_fits_arrays=False, validate_arrays=True) as model:
-            model.validate()
+        # also warn due to the cast_fits_array deprecation
+        with pytest.warns(DeprecationWarning, match="cast_fits_array"):
+            with FitsModel(
+                file_path,
+                strict_validation=True,
+                cast_fits_arrays=False,
+                validate_arrays=True,
+            ) as model:
+                model.validate()
 
     # Wrong dimensions
     hdu = fits.ImageHDU(data=np.ones((4,), dtype=np.float64), name="SCI")
@@ -607,8 +614,10 @@ def test_ndarray_validation(tmp_path):
 
     # Should also be caught by validation
     with pytest.raises(ValidationError, match="Wrong number of dimensions: Expected 2, got 1"):
-        with FitsModel(file_path, strict_validation=True, cast_fits_arrays=False, validate_arrays=True) as model:
-            model.validate()
+        # also warn due to the cast_fits_array deprecation
+        with pytest.warns(DeprecationWarning, match="cast_fits_array"):
+            with FitsModel(file_path, strict_validation=True, cast_fits_arrays=False, validate_arrays=True) as model:
+                model.validate()
 
 
 def test_resave_duplication_bug(tmp_path):