+++ CHANGES.rst
@@ -9,7 +9,8 @@ Bug Fixes
 Changes to API
 --------------
 
-- 
+- Deprecate ``cast_arrays`` argument to ``from_fits_hdu`` and
+  ``cast_fits_arrays`` argument to ``Datamodel.__init__`` [#214]
 
 Other
 -----
@@ -28,6 +29,9 @@ Bug Fixes
 
 - Fix ``rebuild_fits_rec_dtype`` handling of unsigned integer columns
   with shapes [#213]
+
+Changes to API
+--------------
 
 - Sort keyword files used for schema_editor to make output non-arbitrary
   copy schema before merging to avoid schema modification [#227]